+++ e2e/test/jest-mutate-outside-roots/verify/verify.ts
@@ -1,13 +1,8 @@
-import { expectMetrics } from "../../../helpers";
+import { expectMetricsJsonToMatchSnapshot } from "../../../helpers";
 
 describe('jest sut outside roots e2e', () => {
 
   it('should result in the expected mutation score', async () => {
-    await expectMetrics({
-      survived: 1,
-      killed: 3,
-      timeout: 0,
-      noCoverage: 0
-    });
+    await expectMetricsJsonToMatchSnapshot();
   });
 });