+++ e2e/test/jest-node/verify/verify.ts
@@ -2,24 +2,11 @@ import { expect } from 'chai';
 import path from 'path';
 import { promises as fsPromises } from 'fs';
 import { MutationTestResult, TestFileDefinitionDictionary } from 'mutation-testing-report-schema/api';
-import { expectMetricsResult, produceMetrics } from '../../../helpers';
+import { expectMetricsJsonToMatchSnapshot } from '../../../helpers';
 
 describe('After running stryker with test runner jest on test environment "node"', () => {
   it('should report 75% mutation score', async () => {
-    await expectMetricsResult({
-      metrics: produceMetrics({
-        killed: 3,
-        mutationScore: 75,
-        mutationScoreBasedOnCoveredCode: 75,
-        noCoverage: 0,
-        survived: 1,
-        totalCovered: 4,
-        totalDetected: 3,
-        totalMutants: 4,
-        totalUndetected: 1,
-        totalValid: 4
-      })
-    });
+    await expectMetricsJsonToMatchSnapshot();
   });
 
   it('should report test files and test locations', async () => {