+++ e2e/test/jest-react-ts/verify/verify.ts
@@ -1,12 +1,7 @@
-import { expectMetricsJson } from '../../../helpers';
+import { expectMetricsJsonToMatchSnapshot } from '../../../helpers';
 
 describe('After running stryker on jest-react-ts project', () => {
   it('should report expected scores', async () => {
-    await expectMetricsJson({
-      killed: 63,
-      noCoverage: 19,
-      survived: 25,
-      timeout: 5,
-    });
+    await expectMetricsJsonToMatchSnapshot();
   });
 });