+++ packages/core/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.1](https://github.com/stryker-mutator/stryker-js/compare/v5.6.0...v5.6.1) (2022-01-23)
+
+
+### Bug Fixes
+
+* **deps:** remove vulnerability by updating log4js ([#3372](https://github.com/stryker-mutator/stryker-js/issues/3372)) ([69290f2](https://github.com/stryker-mutator/stryker-js/commit/69290f287f30eee9fac96dc32bd1df2f24180b07)), closes [/github.com/log4js-node/log4js-node/blob/master/CHANGELOG.md#640](https://github.com//github.com/log4js-node/log4js-node/blob/master/CHANGELOG.md/issues/640)
+
+
+
+
+
+# [5.6.0](https://github.com/stryker-mutator/stryker-js/compare/v5.5.1...v5.6.0) (2022-01-09)
+
+
+### Bug Fixes
+
+* **report:** dramatically improve rendering performance of HTML report ([ad38c82](https://github.com/stryker-mutator/stryker-js/commit/ad38c8219ab5cd6dc477b67bf3416c9afdfba972))
+
+
+### Features
+
+* **clear-text reporter:** show n/a instead of NaN  ([68c5c51](https://github.com/stryker-mutator/stryker-js/commit/68c5c5183c04ea2cf6b5943996972f4e411e32a9))
+
+
+
+
+
 ## [5.5.1](https://github.com/stryker-mutator/stryker-js/compare/v5.5.0...v5.5.1) (2021-12-02)